simplified webserver logic
