+++ client/ayon_core/tools/tray/webserver/__init__.py
@@ -1,10 +1,13 @@
 from .structures import HostMsgAction
 from .base_routes import RestApiEndpoint
-from .webserver import TrayWebserver
+from .server import find_free_port, WebServerManager
+from .host_console_listener import HostListener
 
 
 __all__ = (
     "HostMsgAction",
     "RestApiEndpoint",
-    "TrayWebserver",
+    "find_free_port",
+    "WebServerManager",
+    "HostListener",
 )
